requirements.txt
streamlit
PyPDF2
langchain
sentence-transformers
faiss-cpu
python-dotenv
langchain-community

